requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
jinja2
scikit-learn
numpy
python-dotenv
